+++ package.json
@@ -11,6 +11,7 @@
     "react-modal": "^3.16.1",
     "react-router-dom": "^6.5.0",
     "react-scripts": "5.0.1",
+    "react-select": "^5.7.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
