+++ src/Components/Header/SigninHeader.jsx
@@ -1,7 +1,9 @@
 import React from "react";
 import "./SigninHeader.css";
+import { useNavigate } from "react-router-dom";
 
 const SigninHeader = () => {
+  const navigate = useNavigate();
   return (
     <div className="header">
       <h1 className="header_title">LAUNDRY</h1>
@@ -10,7 +12,9 @@ const SigninHeader = () => {
           <li className="nav_lists__links">Home</li>
           <li className="nav_lists__links">Pricing</li>
           <li className="nav_lists__links">Career</li>
-          <li className="nav_lists__links color">Sign In</li>
+          <li onClick={() => navigate("/")} className="nav_lists__links color">
+            Sign In
+          </li>
         </ul>
       </nav>
     </div>
