+++ src/Components/HomeHeader/HomeHeader.css
@@ -48,3 +48,29 @@
   background-color: #5861ae;
   color: #ffffff;
 }
+
+.logout {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0 69.9px;
+  font: normal normal normal 18px/19px Avenir;
+  color: #777;
+  border-left: 1px solid rgb(230, 228, 228);
+  cursor: pointer;
+  background-color: red;
+  color: #ffffff;
+  width: 70px;
+  height: 10vh;
+  margin-top: 10vh;
+  margin-left: 84.4%;
+  transition: all 0.5s ease-out;
+}
+
+.home_pic {
+  width: 40px;
+  height: 40px;
+  background-color: #ffffff;
+  margin-right: 10px;
+  border-radius: 50%;
+}
